AR 개선

+++ Assets/TurnTheGameOn/Arrow WayPointer/Scripts/WaypointController.cs
@@ -148,7 +148,7 @@ namespace TurnTheGameOn.ArrowWaypointer{
 
 		#if UNITY_EDITOR
 		//Draws a Gizmo in the scene view window to show the Waypoints
-		public void OnDrawGizmosSelected(int radius) {
+		public void OnDrawGizmosSelected(float radius) {
 			for(var i = 0; i < waypointList.Length; i++){
 				if (waypointList [i] != null) {
 					if (waypointList [i].waypoint != null) {
